// cake_server/src/entities/dto/youzanGood.dto.ts
export class YouzanGood {
	activityPrice: number;
	alias: string;
	id: number;
	imageUrl: string;
	price: number;
	sellPoint: string;
	title: string;
	totalStock: number;
	startSaleNum: number;
	messages: string;
	totalId: number;
	labelViewModel: LabelViewModel;
	productAttributesDefined: boolean;
	preSale: boolean;
	skuSize: number;
	totalSoldNum: number;
	basicTags: any[];
	groupId: number;
}
export class LabelViewModel {
	labelGroupModels: any[];
}